app/backend/requirements.txt
flask
flask-cors
torch
torchvision
timm
numpy
pandas
pillow
matplotlib
scikit-learn
pymongo
pyjwt
bcrypt
python-dotenv
